Removal of manual handling for sorts
