+++ src/Elastic.Clients.Elasticsearch/Types/Mapping/PropertyBase.cs
@@ -13,7 +13,6 @@ public partial interface IProperty
 {
 }
 
-// FUTURE GENERATED
 internal sealed partial class PropertyInterfaceConverter : JsonConverter<IProperty>
 {
 	public override IProperty? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
@@ -42,19 +41,4 @@ internal sealed partial class PropertyInterfaceConverter : JsonConverter<IProper
 	}
 
 	public override void Write(Utf8JsonWriter writer, IProperty value, JsonSerializerOptions options) => throw new NotImplementedException();
-
-	//private static PropertyBase DeserializeVariant(string type, ref Utf8JsonReader reader, JsonSerializerOptions options)
-	//{
-	//	switch (type)
-	//	{
-	//		case "text":
-	//			return JsonSerializer.Deserialize<TextProperty>(ref reader, options);
-
-	//		case "ip":
-	//			return JsonSerializer.Deserialize<IpProperty>(ref reader, options);
-
-	//		default:
-	//			throw new JsonException("Encounted an unknown property type which could not be deserialised.");
-	//	}
-	//}
 }
