+++ src/Elastic.Clients.Elasticsearch/Types/SortCombinations.cs
@@ -1,0 +1,12 @@
+// Licensed to Elasticsearch B.V under one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information.
+
+namespace Elastic.Clients.Elasticsearch;
+
+public partial class SortCombinations
+{
+	public SortCombinations(string field) : base(field)
+	{
+	}
+}
